// apps/web/src/app/api/generated/model/changeCampaignVisibilityRequest.ts
/**
 * CommitQuest API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ChangeCampaignVisibilityRequest {
  visibility: ChangeCampaignVisibilityRequestVisibilityEnum;
}
export enum ChangeCampaignVisibilityRequestVisibilityEnum {
  PRIVATE = 'PRIVATE',
  UNLISTED = 'UNLISTED',
}