// apps/web/src/app/api/generated/model/session.ts
/**
 * CommitQuest API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface Session {
  githubLogin: string;
  displayName: string;
  avatarUrl: string;
  expiresAt: string;
  csrfToken: string;
}