// apps/web/src/app/api/generated/model/savedCampaign.ts
/**
 * CommitQuest API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CampaignProjection } from './campaignProjection';

export interface SavedCampaign {
  id: string;
  projection: CampaignProjection;
  visibility: SavedCampaignVisibilityEnum;
  createdAt: string;
  updatedAt: string;
}
export enum SavedCampaignVisibilityEnum {
  PRIVATE = 'private',
  UNLISTED = 'unlisted',
}