// apps/web/src/app/api/generated/model/savedCampaignExport.ts
/**
 * CommitQuest API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CampaignProjection } from './campaignProjection';

export interface SavedCampaignExport {
  exportSchemaVersion: SavedCampaignExportExportSchemaVersionEnum;
  campaignId: string;
  repository: string;
  title: string;
  visibility: SavedCampaignExportVisibilityEnum;
  projectionSchemaVersion: number;
  mappingAlgorithmVersion: number;
  scoringRulesetVersion: number;
  createdAt: string;
  updatedAt: string;
  projection: CampaignProjection;
}
export enum SavedCampaignExportExportSchemaVersionEnum {
  NUMBER_1 = 1,
}
export enum SavedCampaignExportVisibilityEnum {
  PRIVATE = 'private',
  UNLISTED = 'unlisted',
}